+++ frontend/src/components/form/field.style.ts
@@ -1,0 +1,6 @@
+import { Search } from "@equinor/eds-core-react";
+import styled from "styled-components";
+
+export const SearchFieldInput = styled(Search)`
+  width: 100%;
+`;
