+++ frontend/src/components/form/form.tsx
@@ -1,32 +1,21 @@
-import {
-  Button,
-  DotProgress,
-  TextField as EdsTextField,
-  Icon,
-  InputWrapper,
-  Tooltip,
-} from "@equinor/eds-core-react";
-import { error_filled } from "@equinor/eds-icons";
+import { Button } from "@equinor/eds-core-react";
 import { createFormHook } from "@tanstack/react-form";
-import {
-  ChangeEvent,
-  Dispatch,
-  SetStateAction,
-  useEffect,
-  useState,
-} from "react";
+import { useState } from "react";
 import { toast } from "react-toastify";
-import z from "zod/v4";
 
-import { fieldContext, formContext, useFieldContext } from "#utils/form";
-import { handleValidator, ValidatorProps } from "#utils/validator";
+import { fieldContext, formContext } from "#utils/form";
+import { handleValidator } from "#utils/validator";
+import { SubmitButton } from "./button";
+import {
+  BasicTextFieldProps,
+  CommonTextFieldProps,
+  SearchField,
+  TextField,
+} from "./field";
 import {
   EditableTextFieldFormContainer,
   SearchFieldFormContainer,
-  SearchFieldInput,
 } from "./form.style";
-
-Icon.add({ error_filled });
 
 export type StringObject = { [x: string]: string };
 
@@ -41,14 +30,6 @@ export interface MutationCallbackProps<T> {
   formReset: () => void;
 }
 
-interface BasicTextFieldProps {
-  name: string;
-  label: string;
-  value: string;
-  placeholder?: string;
-  helperText?: string;
-}
-
 interface SetStateFormProps {
   setStateCallback: (value: string) => void;
 }
@@ -56,130 +37,6 @@ interface SetStateFormProps {
 interface MutationFormProps {
   mutationCallback: (props: MutationCallbackProps<StringObject>) => void;
   mutationIsPending: boolean;
-}
-
-export interface CommonTextFieldProps
-  extends BasicTextFieldProps,
-    ValidatorProps {}
-
-export function TextField({
-  label,
-  placeholder,
-  helperText,
-  isReadOnly,
-  toUpperCase,
-  setSubmitDisabled,
-}: {
-  label: string;
-  placeholder?: string;
-  helperText?: string;
-  isReadOnly?: boolean;
-  toUpperCase?: boolean;
-  setSubmitDisabled?: Dispatch<SetStateAction<boolean>>;
-}) {
-  const field = useFieldContext<string>();
-
-  useEffect(() => {
-    if (setSubmitDisabled) {
-      setSubmitDisabled(
-        field.state.meta.isDefaultValue || !field.state.meta.isValid,
-      );
-    }
-  }, [
-    setSubmitDisabled,
-    field.state.meta.isDefaultValue,
-    field.state.meta.isValid,
-  ]);
-
-  return (
-    <InputWrapper helperProps={{ text: helperText }}>
-      <EdsTextField
-        id={field.name}
-        name={field.name}
-        label={label}
-        readOnly={isReadOnly}
-        value={field.state.value}
-        placeholder={placeholder}
-        onBlur={field.handleBlur}
-        onChange={(e: ChangeEvent<HTMLInputElement>) => {
-          let value = e.target.value;
-          if (toUpperCase) {
-            value = value.toUpperCase();
-          }
-          field.handleChange(value);
-        }}
-        {...(!field.state.meta.isValid && {
-          variant: "error",
-          helperIcon: <Icon name="error_filled" title="Error" />,
-          helperText: field.state.meta.errors
-            .map((err: z.ZodError) => err.message)
-            .join(", "),
-        })}
-      />
-    </InputWrapper>
-  );
-}
-
-export function SearchField({
-  placeholder,
-  helperText,
-  toUpperCase,
-}: {
-  placeholder?: string;
-  helperText?: string;
-  toUpperCase?: boolean;
-}) {
-  const field = useFieldContext<string>();
-
-  return (
-    <InputWrapper helperProps={{ text: helperText }}>
-      <SearchFieldInput
-        id={field.name}
-        value={field.state.value}
-        placeholder={placeholder}
-        onBlur={field.handleBlur}
-        onChange={(e) => {
-          let value = e.target.value;
-          if (toUpperCase) {
-            value = value.toUpperCase();
-          }
-          field.handleChange(value);
-        }}
-      />
-    </InputWrapper>
-  );
-}
-
-export function SubmitButton({
-  label,
-  disabled,
-  isPending,
-}: {
-  label: string;
-  disabled?: boolean;
-  isPending?: boolean;
-}) {
-  return (
-    <Tooltip
-      title={
-        disabled
-          ? "Value can be submitted when it has been changed and is valid"
-          : ""
-      }
-    >
-      <Button
-        type="submit"
-        aria-disabled={disabled}
-        onClick={(e) => {
-          if (disabled) {
-            e.preventDefault();
-          }
-        }}
-      >
-        {isPending ? <DotProgress /> : label}
-      </Button>
-    </Tooltip>
-  );
 }
 
 const { useAppForm: useAppFormEditableTextFieldForm } = createFormHook({
